Refactored add-on to subclass existing objects.
Instead of overriding or hooking most functions and crudely maintaining state with global variables, we instead subclass the existing Browser and DataModel objects used by Anki. This has many advantages, including the sharing of state with the rest of the objects they are extending (so we can do things like remove a column as well).