Fix new installations of 1.5
Commit 1295e6d "Left-align columns maintained by this add-on"
introduced another line of initialization which came after the
short-circuit for new installations of the add-on, but is actually
relied on in new and old installations, causing new installations
to go into a crash-loop.  Move it above the short-circuit to
fix that.

Also invert the sense of the conditional to make similar bugs less
likely in the future, and make a few small other improvements while
we're working in this bit of the code: clearer name for a variable,
use `set()` rather than a list for a data structure whose purpose is
membership tests, and collapse a couple of loops into a short list
comprehension.

Fixes #19.