ignore double click on table by default
i.e. make advanced browser more similar to anki

this option was added by Arthur with 7fba8f3 in 2020-03
so that users could edit the table.

the old default "Selectable" according to Arthur's description
in the config.md means "it can be selected. All edition will
be ignored". But if all editing is discarded there's limited
benefit in allowing a user to edit a cell?