no update prompt after installing ankiaddon file
relevant for add-on update checks
AnkiWeb sends branch_index to indicate which branch the user
downloaded. when no branch_index is in the manifest.json file Anki
assumes that you're running the first/oldest version from ankiweb
and hence falsely prompts you to update to a newer add-on version.

today there are five different versions of the AB on ankiweb.
branch_index uses zero-based numbering hence the highest current
branch_index is 4.

This needs to be updated when additional versions only for newer Anki
versions are uploaded to ankiweb.